requirements.txt
google-api-python-client
textblob
nltk
matplotlib
pyyaml
